+++ framework/src/main/java/org/checkerframework/common/wholeprograminference/WholeProgramInferenceJavaParserStorage.java
@@ -1,6 +1,7 @@
 package org.checkerframework.common.wholeprograminference;
 
 import com.github.javaparser.ast.CompilationUnit;
+import com.github.javaparser.ast.Node;
 import com.github.javaparser.ast.NodeList;
 import com.github.javaparser.ast.body.CallableDeclaration;
 import com.github.javaparser.ast.body.ClassOrInterfaceDeclaration;
@@ -14,12 +15,17 @@ import com.github.javaparser.ast.body.RecordDeclaration;
 import com.github.javaparser.ast.body.TypeDeclaration;
 import com.github.javaparser.ast.body.VariableDeclarator;
 import com.github.javaparser.ast.expr.AnnotationExpr;
+import com.github.javaparser.ast.expr.MarkerAnnotationExpr;
+import com.github.javaparser.ast.expr.NormalAnnotationExpr;
 import com.github.javaparser.ast.expr.ObjectCreationExpr;
+import com.github.javaparser.ast.expr.SingleMemberAnnotationExpr;
 import com.github.javaparser.ast.type.ClassOrInterfaceType;
 import com.github.javaparser.ast.type.Type;
 import com.github.javaparser.ast.type.TypeParameter;
 import com.github.javaparser.ast.visitor.CloneVisitor;
+import com.github.javaparser.ast.visitor.VoidVisitor;
 import com.github.javaparser.printer.DefaultPrettyPrinter;
+import com.github.javaparser.printer.DefaultPrettyPrinterVisitor;
 import com.sun.source.tree.ClassTree;
 import com.sun.source.tree.MethodTree;
 import com.sun.source.tree.NewClassTree;
@@ -30,15 +36,18 @@ import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
 import java.io.IOException;
+import java.lang.annotation.Annotation;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import java.util.stream.Collectors;
 import javax.lang.model.element.AnnotationMirror;
 import javax.lang.model.element.Element;
 import javax.lang.model.element.ExecutableElement;
@@ -56,6 +65,7 @@ import org.checkerframework.framework.ajava.AnnotationMirrorToAnnotationExprConv
 import org.checkerframework.framework.ajava.AnnotationTransferVisitor;
 import org.checkerframework.framework.ajava.DefaultJointVisitor;
 import org.checkerframework.framework.ajava.JointJavacJavaParserVisitor;
+import org.checkerframework.framework.qual.InvisibleQualifier;
 import org.checkerframework.framework.qual.TypeUseLocation;
 import org.checkerframework.framework.type.AnnotatedTypeFactory;
 import org.checkerframework.framework.type.AnnotatedTypeMirror;
@@ -104,6 +114,31 @@ public class WholeProgramInferenceJavaParserStorage
 
   /** Whether the {@code -AinferOutputOriginal} option was supplied to the checker. */
   private final boolean inferOutputOriginal;
+
+  /**
+   * Returns the names of all qualifiers that are marked with {@link InvisibleQualifier}, and that
+   * are supported by the given type factory.
+   *
+   * @param atypeFactory a type factory
+   * @return the names of every invisible qualifier supported by {@code atypeFactory}
+   */
+  public static Set<String> getInvisibleQualifierNames(AnnotatedTypeFactory atypeFactory) {
+    return atypeFactory.getSupportedTypeQualifiers().stream()
+        .filter(WholeProgramInferenceJavaParserStorage::isInvisible)
+        .map(Class::getCanonicalName)
+        .collect(Collectors.toSet());
+  }
+
+  /**
+   * Is the definition of the given annotation class annotated with {@link InvisibleQualifier}?
+   *
+   * @param qual an annotation class
+   * @return true iff {@code qual} is meta-annotated with {@link InvisibleQualifier}
+   */
+  public static boolean isInvisible(Class<? extends Annotation> qual) {
+    return Arrays.stream(qual.getAnnotations())
+        .anyMatch(anno -> anno.annotationType() == InvisibleQualifier.class);
+  }
 
   /**
    * Constructs a new {@code WholeProgramInferenceJavaParser} that has not yet inferred any
@@ -782,7 +817,43 @@ public class WholeProgramInferenceJavaParserStorage
       // certain locations. This implementation could be used instead if it's fixed in JavaParser.
       // LexicalPreservingPrinter.print(root.declaration, writer);
 
-      DefaultPrettyPrinter prettyPrinter = new DefaultPrettyPrinter();
+      // Do not print invisible qualifiers, to avoid cluttering the output.
+      Set<String> invisibleQualifierNames = getInvisibleQualifierNames(this.atypeFactory);
+      DefaultPrettyPrinter prettyPrinter =
+          new DefaultPrettyPrinter() {
+            @Override
+            public String print(Node node) {
+              VoidVisitor<Void> visitor =
+                  new DefaultPrettyPrinterVisitor(getConfiguration()) {
+                    @Override
+                    public void visit(final MarkerAnnotationExpr n, final Void arg) {
+                      if (invisibleQualifierNames.contains(n.getName().toString())) {
+                        return;
+                      }
+                      super.visit(n, arg);
+                    }
+
+                    @Override
+                    public void visit(final SingleMemberAnnotationExpr n, final Void arg) {
+                      if (invisibleQualifierNames.contains(n.getName().toString())) {
+                        return;
+                      }
+                      super.visit(n, arg);
+                    }
+
+                    @Override
+                    public void visit(final NormalAnnotationExpr n, final Void arg) {
+                      if (invisibleQualifierNames.contains(n.getName().toString())) {
+                        return;
+                      }
+                      super.visit(n, arg);
+                    }
+                  };
+              node.accept(visitor, null);
+              return visitor.toString();
+            }
+          };
+
       writer.write(prettyPrinter.print(root.compilationUnit));
       writer.close();
     } catch (IOException e) {