+++ framework/src/test/java/org/checkerframework/framework/test/junit/ReflectionTest.java
@@ -10,7 +10,9 @@ import org.junit.runners.Parameterized.Parameters;
 /** Tests the reflection resolution using a simple type system. */
 public class ReflectionTest extends CheckerFrameworkPerDirectoryTest {
 
-  /** @param testFiles the files containing test code, which will be type-checked */
+  /**
+   * @param testFiles the files containing test code, which will be type-checked
+   */
   public ReflectionTest(List<File> testFiles) {
     super(testFiles, ReflectionTestChecker.class, "reflection", "-Anomsgtext");
   }