+++ framework/src/test/java/org/checkerframework/framework/test/junit/ReportTest.java
@@ -7,7 +7,9 @@ import org.junit.runners.Parameterized.Parameters;
 
 public class ReportTest extends CheckerFrameworkPerDirectoryTest {
 
-  /** @param testFiles the files containing test code, which will be type-checked */
+  /**
+   * @param testFiles the files containing test code, which will be type-checked
+   */
   public ReportTest(List<File> testFiles) {
     super(
         testFiles,