+++ framework/src/test/java/org/checkerframework/framework/test/junit/SupportedQualsTest.java
@@ -8,7 +8,9 @@ import org.junit.runners.Parameterized.Parameters;
 
 public class SupportedQualsTest extends CheckerFrameworkPerDirectoryTest {
 
-  /** @param testFiles the files containing test code, which will be type-checked */
+  /**
+   * @param testFiles the files containing test code, which will be type-checked
+   */
   public SupportedQualsTest(List<File> testFiles) {
     super(testFiles, SupportedQualsChecker.class, "simple", "-Anomsgtext");
   }